Once Again From the Top
Correction notice in the Miami Herald: "Last Sunday, The Herald erroneously
reported that original Dolphin Johnny Holmes had been an insurance salesman
in Raleigh, North Carolina, that he had won the New York lottery in 1982 and
lost the money in a land swindle, that he had been charged with vehicular
homicide, but acquitted because his mother said she drove the car, and that
he stated that the funniest thing he ever saw was Flipper spouting water on
George Wilson.  Each of these items was erroneous material published
inadvertently.  He was not an insurance salesman in Raleigh, did not win the
lottery, neither he nor his mother was charged or involved in any way with
vehicular homicide, and he made no comment about Flipper or George Wilson.
The Herald regrets the errors."
		-- "The Progressive", March, 1987